% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{d_geom}
\alias{d_geom}
\title{Returns the distribution value at `x`}
\usage{
d_geom(x, ...)
}
\arguments{
\item{x}{number of trials to simulate}

\item{...}{Optional parameters which are the same as `dgeom`}
}
\value{
A vector of length n with random numbers drawn from the geometric 
distribution with probability p if success = FALSE, or modified geometric
distribution as described above otherwise.
}
\description{
Returns the distribution value at `x`
}
\examples{

dist_geom <- d_geom(4, )
}
