% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{get_msten}
\alias{get_msten}
\title{MSTE-N aggregation function intended for package use only.}
\usage{
get_msten(...)
}
\arguments{
\item{...}{Parameters required for the `rtexp` function.}
}
\value{
Dataframe consisting of `sum` and `max` of `n` `rtexp` random 
variables
}
\description{
MSTE-N aggregation function intended for package use only.
}
\examples{
get_msten(n = 4, rate = -1, b = 6)
}
