% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{inv_texp}
\alias{inv_texp}
\title{Inverse CDF of the truncated exponential distribution intended to be used 
for sampling}
\usage{
inv_texp(x, rate = 1, a = 0, b)
}
\arguments{
\item{x}{Input value, expected to be between 0 and 1}

\item{rate}{Single value specifying the rate.}

\item{a}{Lower truncation bound, if not provided it assumes the default value
of 0.}

\item{b}{Upper truncation bound.}
}
\value{
Numeric vector
}
\description{
Inverse CDF of the truncated exponential distribution intended to be used 
for sampling
}
\examples{
inv_texp(0.25, b = 6)
inv_texp(0.5, rate = -1, a = 2, b = 6)
}
