% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{mdm}
\alias{mdm}
\title{Magnitude, Duration, and Maximum}
\usage{
mdm(x, below = FALSE, duration = 1)
}
\arguments{
\item{x}{A numeric vector. The vector should be standardized so that values
above the threshold are positive and values below are negative. NA values
should be stripped from the vector.}

\item{below}{Boolean indicating if peak above or below threshold should be
calculated. If `TRUE` will calculate MDM values
for events above the threshold. Otherwise will calculate values below
threshold. Default is `TRUE`.}

\item{duration}{Minimum duration required for the MDM of and event to be
calculated. Default is `1`.}
}
\value{
Returns a `data.frame` with trivariate event summaries for all events
found in the vector. See `summarize_event` for further details.
}
\description{
Magnitude, Duration, and Maximum
}
\examples{
observations <- c(5, 4, 3, 4, 2, 1, 1, 2, 4)
threshold <- 2

centered_obs <- observations - threshold
mdm(centered_obs)
mdm(centered_obs, duration = 3)
}
