% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{p_geom}
\alias{p_geom}
\title{Simulate Geometric Distribution with Modified Probability of Success}
\usage{
p_geom(q, ...)
}
\arguments{
\item{q}{probability of the value `q`}

\item{...}{Optional parameters which are the same as `pgeom`}
}
\value{
A vector of length n with random numbers drawn from the geometric 
distribution with probability p if success = FALSE, or modified geometric
distribution as described above otherwise.
}
\description{
Simulate Geometric Distribution with Modified Probability of Success
}
\examples{

sim_geom <- p_geom(n = 100, p = .5)
}
