% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{r_geom}
\alias{r_geom}
\title{Simulate Geometric Distribution with Modified Probability of Success}
\usage{
r_geom(n, p, success = TRUE)
}
\arguments{
\item{n}{number of trials to simulate}

\item{p}{probability of success in each trial (0 <= p < 1)}

\item{success}{logical indicating whether or not to use a modified version of
the geometric distribution that returns ceiling(log(runif())/log(1-p)) 
instead of rgeom() when TRUE. Default is TRUE.}
}
\value{
A vector of length n with random numbers drawn from the geometric 
distribution with probability p if success = FALSE, or modified geometric
distribution as described above otherwise.
}
\description{
Simulate Geometric Distribution with Modified Probability of Success
}
\examples{

sim_geom <- r_geom(n = 100, p = .5, success=TRUE)
}
