% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{rate_estimator}
\alias{rate_estimator}
\title{Function to optimize the `rate_mle()` function. This is a convenience
function  which use the stats `R` function `optimize()`
to numerically find the MLE of the MSTE-N distribution.}
\usage{
rate_estimator(rate_interval = c(-5, 5), optim_tries = 25, b, samp)
}
\arguments{
\item{rate_interval}{Initial interval for estimating the rate parameter.
Defaults to c(-5, 5).}

\item{optim_tries}{Maximum number of optimization attempts. Defaults to 25.}

\item{b}{The upper truncation parameter to pass to `rate_mle()`.}

\item{samp}{The data sample to pass to `rate_mle()`.}
}
\value{
The estimated rate parameter as a single value or a warning message 
if no solution is found within the specified interval and number of 
optimization tries.
}
\description{
Function to optimize the `rate_mle()` function. This is a convenience
function  which use the stats `R` function `optimize()`
to numerically find the MLE of the MSTE-N distribution.
}
\examples{
\dontrun{
rate_estimator(rate_interval = c(-3, 4), optim_tries = 50)
}
}
