% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{rate_mle}
\alias{rate_mle}
\title{Rate Maximum Likelihood Estimation (MLE) Function This function 
calculates the MLE for the rate parameter of the MSTE-N 
distribution. This function has no analytical solution so numerical methods
are required to find an approximate solution.}
\usage{
rate_mle(rate, b, samp)
}
\arguments{
\item{rate}{The rate parameter of a Truncated Exponential distribution.}

\item{b}{The upper truncation parameter.}

\item{samp}{A data frame containing two columns: "N" (typically the duration
of the Trivariate Event) and "x" (the sum of the values measured during the
Trivariate Event).}
}
\value{
The estimate of the rate parameter.
}
\description{
Rate Maximum Likelihood Estimation (MLE) Function This function 
calculates the MLE for the rate parameter of the MSTE-N 
distribution. This function has no analytical solution so numerical methods
are required to find an approximate solution.
}
\examples{
\dontrun{
rate_mle(
  rate = 0.5,
  b = 100, 
  samp = data.frame(
    N = c(rep(20, 10), rep(30, 10)),
    x = rnorm(20)
    )
  )
}
}
