% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{rmsten}
\alias{rmsten}
\title{MSTE-N random vector generator}
\usage{
rmsten(n_dist, rate = 1, a = 0, b, ...)
}
\arguments{
\item{n_dist}{The counting distribution which N from the MSTE-N vector will
be drawn. The number of random variables drawn from this distribution 
determines the number of MSTE-N random vectors drawn.}

\item{rate}{Single value specifying the rate.}

\item{a}{Lower truncation bound, if not provided it assumes the default value
of 0.}

\item{b}{Upper truncation bound.}

\item{...}{The named parameters required for distribution supplied to 
`n_dist`.}
}
\value{
Numeric vector
}
\description{
MSTE-N random vector generator
}
\examples{

rmsten(n_dist = rgeom, rate = -1, a = 2, b = 6, p = 0.25, n = 6)
}
