% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{rtetlg}
\alias{rtetlg}
\title{TETLG random vector generator}
\usage{
rtetlg(n_dist, rate = 1, ...)
}
\arguments{
\item{n_dist}{The counting distribution which N from the TETLG vector will
be drawn. The number of random variables drawn from this distribution 
determines the number of TETLG random vectors drawn.}

\item{rate}{Single value specifying the rate.}

\item{...}{The named parameters required for distribution supplied to 
`n_dist`.}
}
\value{
Numeric vector
}
\description{
TETLG random vector generator
}
\examples{

rtetlg(n_dist = rgeom, rate = 2, p = 0.25, n = 6)
}
