% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{rtexp}
\alias{rtexp}
\title{Truncated exponential random number generator}
\usage{
rtexp(n, rate = 1, a = 0, b)
}
\arguments{
\item{n}{Number of random samples to draw}

\item{rate}{Single value specifying the rate.}

\item{a}{Lower truncation bound, if not provided it assumes the default value
of 0.}

\item{b}{Upper truncation bound.}
}
\value{
Numeric vector
}
\description{
Truncated exponential random number generator
}
\examples{
rtexp(5, b = 6)
rtexp(2, rate = -1, a = 2, b = 6)
}
