% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{summarize_event}
\alias{summarize_event}
\title{Summarize Peak-Over-Threshold event}
\usage{
summarize_event(x, ind)
}
\arguments{
\item{x}{Numeric vector of observations above/below a designated threshold.}

\item{ind}{Index indicating the start and end of single event.}
}
\value{
Returns a data frame summarizing the Trivaraite eveng including the
`sum`, `maximum`, `length`, and `start` and `end` indices for the events.
}
\description{
Summarize Peak-Over-Threshold event
}
\examples{
events <- c(2, 4, 5, -2, -2)
event_index <- c(1, 3)
summarize_event(events, event_index)
}
