% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triVaR-package.R
\docType{package}
\name{triVaR-package}
\alias{triVaR}
\alias{triVaR-package}
\title{triVaR: Trivariate Events}
\description{
Provides helper functions to explore Trivariate Events. This package primarily focuses on the sum and maximum of events with an Exponential or Truncated Exponential distribution.
}
\keyword{internal}
