% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texp.R
\name{r_geom}
\alias{r_geom}
\title{Simulate Geometric Distribution defined in n = 1, 2,...}
\usage{
r_geom(n, ...)
}
\arguments{
\item{n}{Number of trials to simulate}

\item{...}{Optional parameters which are the same as `rgeom`}

\item{p}{Probability of success of one trial}

\item{min_sample}{Integer value to increase the simulated sample size to 
match the duration of episodes with a minimum duration}
}
\value{
A vector of length n with random numbers drawn from the geometric 
distribution with probability p if success = FALSE, or modified geometric
distribution as described above otherwise.
}
\description{
Simulate Geometric Distribution defined in n = 1, 2,...
}
\examples{

sim_geom <- r_geom(n = 100, p = .5, min_sample = 0)
}
